fix: add folder widget for export and moved Upload component in folder component
