+++ src/components/Upload/Upload.scss
