+++ src/components/index.ts
@@ -17,3 +17,4 @@ export { default as Text } from './Text/Text';
 export { default as Toggle } from './Toggle/Toggle';
 export { default as Stepper } from './Stepper/Stepper';
 export { default as Textarea } from './Textarea/Textarea';
+export { default as Upload } from './Upload/Upload';
