+++ src/index.ts
@@ -1,1 +1,2 @@
 export * from './components';
+export * from './widgets';
