+++ src/widgets/index.ts
@@ -1,2 +1,1 @@
 export { default as CollectionCard } from './CollectionCard/CollectionCard';
-export { default as Upload } from './Upload/Upload';
